# task/com.achelos.task.rest-impl/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: "TaSK Framework - Remote API"
  description: "This document defines a RESTful API for executing runs of the TaSK testsuites and accessing their result."
  version: "1.0.0"
  license:
    name: "European Union Public Licence (EUPL) 1.2"
    url: "https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12"
servers:
  - url: '{protocol}://{task_ip}:{task_port}/'
    description: The TaSK REST API.
    variables:
      protocol:
        default: https
        enum:
          - http
          - https
      task_ip:
        default: 127.0.0.1
        description: IP or HostName of TaSK Server, which is served at the location the OpenAPI YAML is served.
      task_port:
        default: '8088'
        description: Port of TaSK Server.
...
paths:
...
  /status/{runId}:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      tags:
        - "Get Status"
      summary: "Retrieve the status of a test suite execution."
      description: "Retrieve the status of the test suite execution specified by the provided run identifier."
      operationId: getStatus
      responses:
        200:
          $ref: '#/components/responses/TestRunStatus'
        400:
          $ref: '#/components/responses/ErrorResponse'
        404:
          $ref: '#/components/responses/ErrorResponse'
        500:
          $ref: '#/components/responses/ErrorResponse'
  /result/{runId}:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      tags:
        - "Get Results"
      summary: "Retrieve the results of a test run."
      description: "Retrieve the results of the test suite execution specified by the provided run identifier."
      operationId: getResult
      responses:
        200:
          $ref: '#/components/responses/TestRunResult'
        400:
          $ref: '#/components/responses/ErrorResponse'
        404:
          $ref: '#/components/responses/ErrorResponse'
        424:
          $ref: '#/components/responses/ErrorResponse'
        500:
          $ref: '#/components/responses/ErrorResponse'
  /result/{runId}/xml:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      tags:
        - "Get Results"
      summary: "Retrieve the XML report of a test suite execution."
      description: "Retrieve XML report of the test suite execution specified by the run identifier."
      operationId: getReportXml
      responses:
        200:
          $ref: '#/components/responses/TestRunReportXml'
        400:
          $ref: '#/components/responses/ErrorResponse'
        404:
          $ref: '#/components/responses/ErrorResponse'
        424:
          $ref: '#/components/responses/ErrorResponse'
        500:
          $ref: '#/components/responses/ErrorResponse'
  /result/{runId}/pdf:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      tags:
        - "Get Results"
      summary: "Retrieve the PDF report of a test suite execution."
      description: "Retrieve PDF report of the test suite execution specified by the run identifier."
      operationId: getReportPdf
      responses:
        200:
          $ref: '#/components/responses/TestRunReportPdf'
        400:
          $ref: '#/components/responses/ErrorResponse'
        404:
          $ref: '#/components/responses/ErrorResponse'
        424:
          $ref: '#/components/responses/ErrorResponse'
        500:
          $ref: '#/components/responses/ErrorResponse'
components:
  parameters:
    RunId:
      name: runId
      description: "Identifier of a Testrun/Test Suite Execution."
      in: path
      required: true
      schema:
        type: string
        example: a4c6ffd34448adef23
...
  responses:
...
    TestRunStatus:
      description: "OK"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Status'
    TestRunResult:
      description: "OK"
      content:
        application/octet-string:
          schema:
            $ref: '#/components/schemas/ResultArchive'
      headers:
        Content-Disposition:
          schema:
            type: string
            example: attachment; filename="TestReport_62b35581-d9d8-479d-9eb2-7f9995b495ad.zip"
    TestRunReportXml:
      description: "OK"
      content:
        application/xml:
          schema:
            $ref: '#/components/schemas/XmlReport'
      headers:
        Content-Disposition:
          schema:
            type: string
            example: attachment; filename="TestReport_62b35581-d9d8-479d-9eb2-7f9995b495ad.xml"
    TestRunReportPdf:
      description: "OK"
      content:
        application/pdf:
          schema:
            $ref: '#/components/schemas/PdfReport'
      headers:
        Content-Disposition:
          schema:
            type: string
            example: attachment; filename="TestReport_62b35581-d9d8-479d-9eb2-7f9995b495ad.pdf"
    ErrorResponse:
      description: "ErrorMessage"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
        statuscode:
          type: string
        errormessage:
          type: string
        path:
          type: string
    ResultArchive:
      type: string
      format: binary
    PdfReport:
      type: string
      format: binary
...
    Status:
      type: object
      properties:
        status:
          type: string
          enum:
            - Scheduled
            - Running
            - Aborted
            - Executed
        downloadURI:
          type: string
          format: url
    XmlReport:
      type: object
      properties: {}
      example: See schema file TaSKReport.xsd.